add to db

+++ src/main/java/com/example/currencyexchange/repository/AuditInfoRepository.java
@@ -1,11 +1,11 @@
 package com.example.currencyexchange.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.example.currencyexchange.models.AuditInfo;
 import org.springframework.stereotype.Repository;
 
+import com.example.currencyexchange.models.AuditInfo;
+
 @Repository
-public interface AuditInfoRepository extends JpaRepository<AuditInfo, Long>{
-    
-}
+public interface AuditInfoRepository extends JpaRepository<AuditInfo, Integer>{
+        
+}
